+++ test/test_basic_logic.py
@@ -114,6 +114,23 @@ class TestBasicClient(object):
             expected_data_frame_pad_length=0
         )
 
+    def test_sending_data_in_memoryview(self):
+        """
+        Support memoryview for sending data.
+        """
+        c = h2.connection.H2Connection()
+        c.initiate_connection()
+        c.send_headers(1, self.example_request_headers)
+
+        # Clear the data, then send some data.
+        c.clear_outbound_data_buffer()
+        events = c.send_data(1, memoryview(b'some data'))
+        assert not events
+        data_to_send = c.data_to_send()
+        assert (
+            data_to_send == b'\x00\x00\t\x00\x00\x00\x00\x00\x01some data'
+        )
+
     def test_sending_data_with_padding(self):
         """
         Single data frames with padding are encoded correctly.